[mod/#146] StoreOptionState 분리
