+++ feature/store/src/main/java/com/napzak/market/store/store/state/StoreUiState.kt
@@ -2,9 +2,6 @@ package com.napzak.market.store.store.state
 
 import androidx.compose.runtime.Immutable
 import com.napzak.market.common.state.UiState
-import com.napzak.market.common.type.MarketTab
-import com.napzak.market.common.type.SortType
-import com.napzak.market.genre.model.Genre
 import com.napzak.market.product.model.Product
 import com.napzak.market.store.model.StoreDetail
 
@@ -27,13 +24,3 @@ data class StoreUiState(
             else -> UiState.Empty
         }
 }
-
-@Immutable
-data class StoreOptionState(
-    val selectedTab: MarketTab = MarketTab.SELL,
-    val filteredGenres: List<Genre> = emptyList<Genre>(),
-    val initGenreItems: List<Genre> = emptyList<Genre>(),
-    val genreSearchResultItems: List<Genre> = emptyList<Genre>(),
-    val isOnSale: Boolean = false,
-    val sortOption: SortType = SortType.RECENT,
-)
